+++ enclone_visual/src/bin/test_vis.rs
@@ -4,6 +4,7 @@
 //
 // If you run with the single argument UPDATE, failing results will be replaced.
 
+use enclone_visual::testsuite::TESTS;
 use perf_stats::*;
 use pretty_trace::*;
 use std::env;
@@ -33,7 +34,7 @@ fn main() {
     }
     print!("{}", strme(&o.stdout));
     let mut fail = false;
-    for i in 1..=4 {
+    for i in 1..=TESTS.len() {
         let (mut image_old, mut image_new) = (Vec::<u8>::new(), Vec::<u8>::new());
         let old_file = format!("enclone_visual/regression_images/test{}.png", i);
         let mut f = File::open(&old_file).unwrap();
