+++ enclone_visual/src/lib.rs
@@ -33,9 +33,13 @@ pub mod geometry;
 pub mod gui;
 pub mod gui_structures;
 pub mod messages;
+pub mod process_messages;
 pub mod style;
 pub mod svg_to_geometry;
+pub mod testsuite;
 pub mod update_restart;
+
+type MsgFn = fn(Result<(), String>) -> messages::Message;
 
 async fn noop() -> Result<(), String> {
     // Increasing this time to 2000ms will prevent the screen from going dark on initialization
